+++ app/src/main/java/com/billyluisneedham/whatsinthefridge/data/source/FoodTagDataSource.kt
@@ -1,8 +1,8 @@
-package com.billyluisneedham.whatsinthefridge.source
+package com.billyluisneedham.whatsinthefridge.data.source
 
-import com.billyluisneedham.whatsinthefridge.model.FoodTag
-import com.billyluisneedham.whatsinthefridge.model.FoodType
-import com.billyluisneedham.whatsinthefridge.model.TagType
+import com.billyluisneedham.whatsinthefridge.domain.model.FoodTag
+import com.billyluisneedham.whatsinthefridge.domain.model.FoodType
+import com.billyluisneedham.whatsinthefridge.domain.model.TagType
 
 interface FoodTagDataSource {
