+++ app/src/main/java/com/billyluisneedham/whatsinthefridge/data/source/TagTypeDataSource.kt
@@ -1,6 +1,6 @@
-package com.billyluisneedham.whatsinthefridge.source
+package com.billyluisneedham.whatsinthefridge.data.source
 
-import com.billyluisneedham.whatsinthefridge.model.TagType
+import com.billyluisneedham.whatsinthefridge.domain.model.TagType
 
 interface TagTypeDataSource {
