+++ app/src/main/java/com/billyluisneedham/whatsinthefridge/domain/model/FoodInstance.kt
@@ -1,4 +1,4 @@
-package com.billyluisneedham.whatsinthefridge.model
+package com.billyluisneedham.whatsinthefridge.domain.model
 
 data class FoodInstance(
     val foodInstanceId: Long = 0,
