+++ app/src/main/java/com/billyluisneedham/whatsinthefridge/domain/model/FoodType.kt
@@ -1,4 +1,4 @@
-package com.billyluisneedham.whatsinthefridge.model
+package com.billyluisneedham.whatsinthefridge.domain.model
 
 data class FoodType(
     val foodTypeId: Long = 0,
