+++ app/src/main/java/com/billyluisneedham/whatsinthefridge/ui/foodlist/FoodCardComponents.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.billyluisneedham.whatsinthefridge.R
 import com.billyluisneedham.whatsinthefridge.mocks.FoodMocks
-import com.billyluisneedham.whatsinthefridge.model.FoodInstance
+import com.billyluisneedham.whatsinthefridge.domain.model.FoodInstance
 import com.billyluisneedham.whatsinthefridge.ui.theme.WhatsInTheFridgeTheme
 import com.billyluisneedham.whatsinthefridge.util.timedisplayer.TimeDisplayer
 import com.billyluisneedham.whatsinthefridge.util.timedisplayer.TimeDisplayerImpl
